Rename and move tabs to different directory
