+++ src/gui/tabs/global_registration_tab.py
