+++ src/gui/tabs/local_registration_tab.py
