+++ src/gui/tabs/multi_scale_registration_tab.py
