+++ src/gui/tabs/rasterizer_tab.py
