+++ src/gui/windows/main_window.py
@@ -3,17 +3,18 @@ import os
 import numpy as np
 from PyQt5.QtCore import QThread, Qt
 from PyQt5.QtWidgets import QMainWindow, QSplitter, QWidget, QGroupBox, QVBoxLayout, \
-    QTabWidget, QSizePolicy, QErrorMessage, QMessageBox, QProgressDialog, QLabel
+    QTabWidget, QSizePolicy, QErrorMessage, QMessageBox, QProgressDialog
 
-from src.gui.widgets.cache_tab_widget import CacheTab
-from src.gui.widgets.global_registration_widget import GlobalRegistrationTab
-from src.gui.widgets.input_tab_widget import InputTab
-from src.gui.widgets.local_registration_widget import LocalRegistrationTab
-from src.gui.widgets.merger_widget import MergerWidget
-from src.gui.widgets.multi_scale_registration_widget import MultiScaleRegistrationTab
-from src.gui.widgets.rasterizer_tab_widget import RasterizerTab
+from src.gui.tabs.cache_tab import CacheTab
+from src.gui.tabs.evaluation_tab import EvaluationTab
+from src.gui.tabs.global_registration_tab import GlobalRegistrationTab
+from src.gui.tabs.input_tab import InputTab
+from src.gui.tabs.local_registration_tab import LocalRegistrationTab
+from src.gui.tabs.merger_tab import MergeTab
+from src.gui.tabs.multi_scale_registration_tab import MultiScaleRegistrationTab
+from src.gui.tabs.rasterizer_tab import RasterizerTab
 from src.gui.widgets.transformation_widget import Transformation3DPicker
-from src.gui.widgets.visualizer_widget import VisualizerWidget
+from src.gui.tabs.visualizer_tab import VisualizerTab
 from src.gui.windows.image_viewer_window import RasterImageViewer
 from src.gui.windows.open3d_window import Open3DWindow
 from src.gui.workers.qt_fgr_registrator import FGRRegistrator
@@ -101,9 +102,9 @@ class RegistrationMainWindow(QMainWindow):
         self.input_tab = InputTab(self.input_dir)
         self.cache_tab = CacheTab(self.cache_dir)
         self.transformation_picker = Transformation3DPicker()
-        self.visualizer_widget = VisualizerWidget()
+        self.visualizer_widget = VisualizerTab()
         self.rasterizer_tab = RasterizerTab()
-        self.merger_widget = MergerWidget(self.output_dir, self.input_dir)
+        self.merger_widget = MergeTab(self.output_dir, self.input_dir)
 
         self.transformation_picker.transformation_matrix_changed.connect(self.update_point_clouds)
         self.input_tab.result_signal.connect(self.handle_result)
@@ -138,9 +139,13 @@ class RegistrationMainWindow(QMainWindow):
         multi_scale_registration_widget = MultiScaleRegistrationTab(self.input_dir)
         multi_scale_registration_widget.signal_do_registration.connect(self.do_multi_scale_registration)
 
+        evaluator_widget = EvaluationTab()
+
         registration_tab.addTab(global_registration_widget, "Global Registration")
         registration_tab.addTab(local_registration_widget, "Local Registration")
         registration_tab.addTab(multi_scale_registration_widget, "Multi-scale")
+        registration_tab.addTab(multi_scale_registration_widget, "Multi-scale")
+        registration_tab.addTab(evaluator_widget, "Evaluation")
         layout.addWidget(registration_tab)
 
     # Event Handlers
